pyproject.toml
[tool.poetry]
name = "chocothecoo"
version = "0.1.0"
description = ""
authors = ["sodasora <awo3sr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
requests = "^2.31.0"
djangorestframework-simplejwt = "^5.2.2"
python-dotenv = "^1.0.0"
django-cors-headers = "^4.0.0"
pillow = "^9.5.0"
pycryptodome = "^3.18.0"
black = "^23.3.0"
channels = {version = ">=4.0.0,<4.1.0", extras = ["daphne"]}
redis = "^4.5.5"
channels-redis = "^4.1.0"
django-environ = "^0.10.0"
psycopg2 = "^2.9.6"
django-storages = "^1.13.2"
boto3 = "^1.26.160"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
